// src/interfaces/ITestData.ts
export interface ITestData {
    questionID: number;
    question_Type: string;
    pglevel: number;
    answer_selection: number;
    answer_first_type?: string;
    answer_first_content?: string;
    answer_second_type?: string;
    answer_second_content?: string;
    answer_third_type?: string;
    answer_third_content?: string;
}